[USUARIOS] Adicionando verificadores de cadastro
