--- frontend/src/app/features/teste/pages/cadastro/cadastro.component.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="cadastro.component.css">
-    <title>eCapibaribe - Cadastro</title>
-</head>
-<body>
-    <h2 class="brand-logo center" style="font-weight: 600;" >Cadastro</h2>
-    <div class="row">
-        <form class="col s10 offset-s2" (ngSubmit)="onSubmit()">
-            <div class="row">
-                <div class="input-field col s5 form-group">
-                    <label for="nome">Nome*</label>
-                    <input id="nome" type="text" class="form-control" name="nome" [(ngModel)]="user.nome" required>
-                </div>
-                <div class="input-field col s5 form-group">
-                    <label for="sobrenome">Sobrenome*</label>
-                    <input id="sobrenome" type="text" class="form-control" name="sobrenome" [(ngModel)]="user.sobrenome" required>
-                </div>
-            </div>
-            <div class="row">
-                <div class="input-field col s10 form-group">
-                    <label for="cpf">CPF*</label>
-                    <input id="cpf" type="text" class="form-control" name="cpf" [(ngModel)]="user.cpf" required>
-                </div>
-            </div>
-            <div class="row">
-                <div class="input-field col s5 form-group">
-                    <label for="telefone">Telefone</label>
-                    <input id="telefone" type="number" class="form-control" name="telefone" [(ngModel)]="user.telefone">
-                </div>
-                <div class="input-field col s5 form-group">
-                    <label for="dataNascimento">Data de Nascimento</label>
-                    <input id="dataNascimento" type="text" class="form-control" name="dataNascimento" [(ngModel)]="user.dataNascimento">
-                </div>
-            </div>
-            <div class="row">
-                <div class="input-field col s10 form-group">
-                    <label for="email">Email*</label>
-                    <input id="email" type="email" class="form-control" name="email" [(ngModel)]="user.email" required>
-                </div>
-            </div>
-            <div class="row">
-                <div class="input-field col s10 form-group">
-                    <label for="senha">Senha*</label>
-                    <input id="senha" type="password" class="form-control" name="senha" [(ngModel)]="user.senha" required>
-                </div>
-            </div>
-            <button class="btn waves-effect waves-light offset-s2" type="submit" value="enviar">Enviar
-            </button>
-        </form>
-    </div>
-</body>
-</html>
