+++ frontend/src/app/features/users/pages/cadastro/cadastro.component.html
@@ -1,0 +1,96 @@
+<!DOCTYPE html>
+<html lang="pt">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="cadastro.component.css">
+    <link rel="stylesheet" href="https://unpkg.com/bootstrap@4.6.0/dist/css/bootstrap.min.css" />
+    <title>eCapibaribe - Cadastro</title>
+</head>
+<body>
+    <h2 class="brand-logo center" style="font-weight: 600;" >Cadastro</h2>
+    <div class="row">
+        <form class="col s10 offset-s2" #registerForm="ngForm" (ngSubmit)="registerForm.form.valid && onSubmit()" novalidate>
+            <div class="row">
+                <div class="input-field col s5 form-group">
+                    <label for="nome">Nome*</label>
+                    <input id="nome" type="text" class="form-control" name="nome" [(ngModel)]="user.nome" #nome="ngModel" required>
+                    <span
+                      class="text-danger"
+                      *ngIf="(nome.touched || registerForm.submitted) && nome.errors?.['required']"
+                    >
+                      Campo obrigatório
+                    </span>
+                </div>
+                <div class="input-field col s5 form-group">
+                    <label for="sobrenome">Sobrenome*</label>
+                    <input id="sobrenome" type="text" class="form-control" name="sobrenome" [(ngModel)]="user.sobrenome" #sobrenome="ngModel" required>
+                    <span
+                      class="text-danger"
+                      *ngIf="(sobrenome.touched || registerForm.submitted) && sobrenome.errors?.['required']"
+                    >
+                      Campo obrigatório
+                    </span>
+                </div>
+            </div>
+            <div class="row">
+                <div class="input-field col s10 form-group">
+                    <label for="cpf">CPF*</label>
+                    <input id="cpf" type="text" class="form-control" name="cpf" [(ngModel)]="user.cpf" #cpf="ngModel" required>
+                    <span
+                      class="text-danger"
+                      *ngIf="(cpf.touched || registerForm.submitted) && cpf.errors?.['required']"
+                    >
+                      Campo obrigatório
+                    </span>
+                </div>
+            </div>
+            <div class="row">
+                <div class="input-field col s5 form-group">
+                    <label for="telefone">Telefone</label>
+                    <input id="telefone" type="number" class="form-control" name="telefone" [(ngModel)]="user.telefone">
+                </div>
+                <div class="input-field col s5 form-group">
+                    <label for="dataNascimento">Data de Nascimento</label>
+                    <input id="dataNascimento" type="text" class="form-control" name="dataNascimento" [(ngModel)]="user.dataNascimento">
+                </div>
+            </div>
+            <div class="row">
+                <div class="input-field col s10 form-group">
+                    <label for="email">Email*</label>
+                    <input id="email" type="email" class="form-control" name="email" [(ngModel)]="user.email" #email="ngModel" required>
+                    <span
+                      class="text-danger"
+                      *ngIf="(email.touched || registerForm.submitted) && email.errors?.['required']"
+                    >
+                      Campo obrigatório
+                    </span>
+                    <span
+                        class="text-danger"
+                        *ngIf="registerForm.submitted && patternValidatorEmail(user.email)">
+                        Email inválido
+                    </span>
+                </div>
+            </div>
+            <div class="row">
+                <div class="input-field col s10 form-group">
+                    <label for="senha">Senha*</label>
+                    <input id="senha" type="password" class="form-control" name="senha" [(ngModel)]="user.senha" #senha="ngModel" required>
+                    <span
+                        class="text-danger"
+                        *ngIf="(senha.touched || registerForm.submitted) && senha.errors?.['required']">
+                        Campo obrigatório
+                    </span>
+                    <span
+                        class="text-danger"
+                        *ngIf="registerForm.submitted && patternValidator(user.senha)">
+                        A senha deve ter pelo menos 6 caracteres, 1 caractere especial, 1 letra maiúscula, 1 letra minúscula e 1 dígito.
+                    </span>
+                </div>
+            </div>
+            <button class="btn waves-effect waves-light offset-s2" type="submit" value="enviar">Enviar
+            </button>
+        </form>
+    </div>
+</body>
+</html>
